backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
streamlit
yfinance
plotly
pandas
redis
streamlit-autorefresh
textblob
langchain
langchain-core
langchain-google-genai
langgraph
python-dotenv
pytz
requests
